fastjet/internal/predicates: change over to mat
