+++ fastjet/internal/predicates/orientation_test.go
@@ -6,8 +6,6 @@ package predicates
 
 import (
 	"testing"
-
-	"gonum.org/v1/gonum/mat"
 )
 
 func TestOrientation(t *testing.T) {
@@ -133,34 +131,4 @@ func BenchmarkRobustOrientation(b *testing.B) {
 			robustOrientation(setBig(test.x1), setBig(test.y1), setBig(test.x2), setBig(test.y2), setBig(test.x), setBig(test.y))
 		}
 	}
-}
-
-// matOrientation determines the orientation using the mat package.
-//
-// It first computes the conditional number of the matrix. When the condition number
-// is higher than the Condition Tolerance, then we assume the matrix is singular and
-// the determinant is 0. If the determinant is not 0 the sign of the determinant is computed.
-//  | x1 y1 1 |
-//  | x2 y2 1 |
-//  | x  y  1 |
-// FIXME once LU.Cond() is exported do the factorization here to improve performance
-func matOrientation(x1, y1, x2, y2, x, y float64) OrientationKind {
-	if (x1 == x2 && x2 == x) || (y1 == y2 && y2 == y) {
-		// points are horizontally or vertically aligned
-		return Colinear
-	}
-	m := mat.NewDense(3, 3, []float64{x1, y1, 1, x2, y2, 1, x, y, 1})
-	cond := mat.Cond(m, 1)
-	if cond > mat.ConditionTolerance {
-		return Colinear
-	}
-	// Since only the sign is needed LogDet achieves the result in faster time.
-	_, sign := mat.LogDet(m)
-	switch sign {
-	case 1:
-		return CCW
-	case -1:
-		return CW
-	}
-	return IndeterminateOrientation
 }
